feature: extracted store related code
