+++ app/code/Deity/MagentoApi/etc/extension_attributes.xml
@@ -42,12 +42,4 @@
     <extension_attributes for="Magento\Sales\Api\Data\OrderPaymentInterface">
         <attribute code="method_name" type="string" />
     </extension_attributes>
-    <extension_attributes for="Magento\Store\Api\Data\StoreConfigInterface">
-        <attribute code="optional_post_codes" type="mixed" />
-        <attribute code="min_password_length" type="int" />
-        <attribute code="min_password_char_class" type="int" />
-    </extension_attributes>
-    <extension_attributes for="Magento\Store\Api\Data\StoreInterface">
-        <attribute code="is_active" type="int" />
-    </extension_attributes>
 </config>
