+++ app/code/Deity/MagentoApi/etc/webapi_rest/di.xml
@@ -74,12 +74,4 @@
     <type name="Magento\Quote\Model\QuoteManagement">
         <plugin name="deityMagentoApiQuoteManagement" type="Deity\MagentoApi\Plugin\Quote\Model\QuoteManagement" />
     </type>
-
-    <type name="Magento\Store\Api\StoreConfigManagerInterface">
-        <plugin name="deityMagentoApiStoreConfigManager" type="Deity\MagentoApi\Plugin\Store\Api\StoreConfigManager" />
-    </type>
-
-    <type name="Magento\Store\Api\StoreRepositoryInterface">
-        <plugin name="deityMagentoApiStoreRepositoryPlugin" type="Deity\MagentoApi\Plugin\Store\Api\StoreRepository" />
-    </type>
 </config>
