+++ app/code/Deity/Store/README.md
@@ -1,0 +1,1 @@
+### Deity Store Module
