+++ app/code/Deity/Store/composer.json
@@ -1,0 +1,22 @@
+{
+    "name": "deity/falcon-magento-store",
+    "description": "Deity Magento Store integration",
+    "type": "magento2-module",
+    "version": "1.0.0",
+    "require": {
+        "magento/framework": "*",
+        "magento/module-store" : "*",
+        "php": "^7.0"
+    },
+    "license": [
+        "OSL-3.0"
+    ],
+    "autoload": {
+        "files": [
+            "registration.php"
+        ],
+        "psr-4": {
+            "Deity\\Store\\": ""
+        }
+    }
+}
