+++ app/code/Deity/Store/registration.php
@@ -1,0 +1,8 @@
+<?php
+declare(strict_types=1);
+
+\Magento\Framework\Component\ComponentRegistrar::register(
+    \Magento\Framework\Component\ComponentRegistrar::MODULE,
+    "Deity_Store",
+    __DIR__
+);
